add transfer with local signature

+++ src/main/java/io/jafka/jeos/core/common/SignArg.java
@@ -19,6 +19,6 @@ public class SignArg {
     LocalDateTime headBlockTime;
     String chainId;
     //
-    Long expiredSecond;
+    int expiredSecond;
 
 }
